fix: fix segfault, and move links function in link.c
