+++ header/link.h
@@ -1,0 +1,8 @@
+#ifndef LINK_H
+# define LINK_H
+
+# include "lem_in.h"
+
+void set_link_in_rooms(t_lem_in *lem_in, char *room1, char *room2);
+
+#endif
