+++ src/room.c
@@ -1,46 +1,5 @@
 #include "lem_in.h"
 
-static void free_old_array(char **arr, size_t size)
-{
-	size_t i;
-
-	i = 0;
-	while (i < size)
-	{
-		free(arr[i]);
-		i++;
-	}
-	free(arr);
-}
-
-static void *add_link(t_array *links, char *link)
-{
-	char *dup_link;
-
-	dup_link = ft_strdup(link);
-	if (!dup_link)
-		return (NULL);
-	links->size += 1;
-	char **tmp = (char **)ft_realloc(links->arr, sizeof(char *), links->size + 1);
-	if (!tmp)
-		return (NULL);
-	free_old_array(links->arr, links->size - 1);
-	links->arr = tmp;
-	links->arr[links->size - 1] = dup_link;
-	return (links);
-}
-
-void set_link_in_rooms(t_lem_in *lem_in, char *room1, char *room2)
-{
-    t_room *room;
-
-	room = get_room(lem_in, room1);
-	if (!room)
-		fatal_errors_handler(lem_in, "Room does not exist.\n");
-	if (!add_link(&room->links, room2))
-		fatal_errors_handler(lem_in, "Malloc error.\n");
-}
-
 void set_room(t_room *room, int id, char *name, t_vector pos)
 {
     room->id = id;
